Refactor autosuggest keydown controller
