+++ src/autosuggest/options-controller.ts
@@ -10,8 +10,6 @@ import {
   HighlightedOptionState,
   useHighlightedOption,
 } from '../internal/components/options-list/utils/use-highlight-option';
-import { KeyCode } from '../internal/keycode';
-import { BaseKeyDetail } from '../internal/events';
 
 type Options = AutosuggestProps.Options;
 
@@ -31,7 +29,7 @@ export interface AutosuggestItemsState extends HighlightedOptionState<Autosugges
 
 export interface AutosuggestItemsHandlers extends HighlightedOptionHandlers<AutosuggestItem> {
   setShowAll(value: boolean): void;
-  interceptKeyDown(event: CustomEvent<BaseKeyDetail>): boolean;
+  selectHighlightedOptionWithKeyboard(): boolean;
   highlightVisibleOptionWithMouse(index: number): void;
   selectVisibleOptionWithMouse(index: number): void;
 }
@@ -71,26 +69,12 @@ export const useAutosuggestItems = ({
     isHighlightable,
   });
 
-  const interceptKeyDown = (event: CustomEvent<BaseKeyDetail>): boolean => {
-    switch (event.detail.keyCode) {
-      case KeyCode.down: {
-        highlightedOptionHandlers.moveHighlightWithKeyboard(1);
-        return true;
-      }
-      case KeyCode.up: {
-        highlightedOptionHandlers.moveHighlightWithKeyboard(-1);
-        return true;
-      }
-      case KeyCode.enter: {
-        if (highlightedOptionState.highlightedOption && isInteractive(highlightedOptionState.highlightedOption)) {
-          onSelectItem(highlightedOptionState.highlightedOption);
-          return true;
-        }
-        return false;
-      }
-      default:
-        return false;
+  const selectHighlightedOptionWithKeyboard = () => {
+    if (highlightedOptionState.highlightedOption && isInteractive(highlightedOptionState.highlightedOption)) {
+      onSelectItem(highlightedOptionState.highlightedOption);
+      return true;
     }
+    return false;
   };
 
   const highlightVisibleOptionWithMouse = (index: number) => {
@@ -110,7 +94,7 @@ export const useAutosuggestItems = ({
     {
       ...highlightedOptionHandlers,
       setShowAll,
-      interceptKeyDown,
+      selectHighlightedOptionWithKeyboard,
       highlightVisibleOptionWithMouse,
       selectVisibleOptionWithMouse,
     },
